# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API backend",
        "description": "Documentação da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Lista todos os pedidos do usuário autenticado.",
                "operationId": "e848822b988b4e4cf96af0f71f29581c",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pedidos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pedido"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Cria um novo pedido.",
                "operationId": "36884e796fd6f470919017f7ca543cd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_pedido",
                                    "descricao",
                                    "valor",
                                    "prazo_entrega"
                                ],
                                "properties": {
                                    "data_pedido": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-22"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Desenvolvimento de site"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "prazo_entrega": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Mostra um pedido específico.",
                "operationId": "41f9fd012dcc6f1915be1a7372c7d87b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Atualiza um pedido.",
                "operationId": "c552afd071c0184bb7ded11c11c3bb41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "prazo_entrega": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "progresso": {
                                        "type": "string"
                                    },
                                    "comentarios": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão para editar esse pedido"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}/aceitar": {
            "post": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Aceita um pedido.",
                "operationId": "6c26d0cd6f03df326602470b18487aa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido aceito com sucesso"
                    },
                    "400": {
                        "description": "Pedido já foi aceito"
                    },
                    "403": {
                        "description": "Sem permissão para aceitar"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}/iniciar": {
            "post": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Inicia o desenvolvimento de um pedido.",
                "operationId": "2bdb22af94790d019b4422aba68939b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido iniciado com sucesso"
                    },
                    "400": {
                        "description": "Pedido já está em andamento"
                    },
                    "403": {
                        "description": "Sem permissão para iniciar"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}/cancelar": {
            "post": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Cancela um pedido.",
                "operationId": "59b339e29af50124ecf4d10100bffd8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido cancelado com sucesso"
                    },
                    "400": {
                        "description": "Pedido já finalizado"
                    },
                    "403": {
                        "description": "Sem permissão para cancelar"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}/concluir": {
            "post": {
                "tags": [
                    "Pedido"
                ],
                "summary": "Conclui um pedido.",
                "operationId": "c911e1474f9e0dee2115f613f9fedd15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido concluído com sucesso"
                    },
                    "400": {
                        "description": "Pedido já finalizado"
                    },
                    "403": {
                        "description": "Sem permissão para concluir"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pedido": {
                "title": "Pedido",
                "required": [
                    "id",
                    "data_pedido",
                    "descricao",
                    "valor",
                    "prazo",
                    "id_empresa"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "data_pedido": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-22"
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Desenvolvimento de site"
                    },
                    "valor": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "prazo": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-22"
                    },
                    "id_empresa": {
                        "type": "integer",
                        "example": 1
                    },
                    "desenvolvedor_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "status": {
                        "type": "array",
                        "items": {
                            "required": [
                                "status",
                                "data_status"
                            ],
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "aguardando"
                                },
                                "data_status": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2025-05-22T14:55:00Z"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "pagamento": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "id_pedido": {
                                "type": "integer",
                                "example": 1
                            },
                            "status": {
                                "type": "string",
                                "example": "pago"
                            },
                            "valor_pago": {
                                "type": "number",
                                "format": "float",
                                "example": 1500
                            },
                            "data_pagamento": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-05-23T10:00:00Z"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": " * @OA\\SecurityScheme(\n * securityScheme=\"sanctum\",\n * type=\"http\",\n * scheme=\"bearer\",\n * bearerFormat=\"Sanctum\"\n * )\n ",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}